+++ src/extension.ts
@@ -10,8 +10,8 @@ import { registerAppServiceExtensionVariables } from 'vscode-azureappservice';
 import { AzExtTreeDataProvider, callWithTelemetryAndErrorHandling, createApiProvider, createAzExtOutputChannel, createExperimentationService, IActionContext, registerUIExtensionVariables } from 'vscode-azureextensionui';
 import { AzureExtensionApi, AzureExtensionApiProvider } from 'vscode-azureextensionui/api';
 import { revealTreeItem } from './commands/api/revealTreeItem';
-import { contentScheme } from './commands/github/logs/GitHubLogContentProvider';
-import GitHubLogFoldingProvider from './commands/github/logs/GitHubLogFoldingProvider';
+import { contentScheme } from './commands/github/jobLogs/GitHubLogContentProvider';
+import GitHubLogFoldingProvider from './commands/github/jobLogs/GitHubLogFoldingProvider';
 import { registerCommands } from './commands/registerCommands';
 import { githubAuthProviderId, githubScopes } from './constants';
 import { ext } from './extensionVariables';
